blog/requirements.txt
Django==2.0.5
django-contrib-comments==1.8.0
django-cors-headers==2.2.0
django-polls==0.1
djangorestframework==3.8.2
mysqlclient==1.3.12
pytz==2018.4
selenium==3.12.0
